Ensure cloning Vundle repo as the right user.
Since a vulnerability was discovered, Git intentionally skips cloning of repos into parent folders owned by other users (actually not, it's more of an environment variables thing but lets just assume it for the sake of the commit message. I'm documenting my project, not Git daddy issues. Deal with it).

To know more:
* https://stackoverflow.com/questions/72299489/unsafe-repository-owned-by-someone-else-ok-but-who
* https://stackoverflow.com/questions/71901632/fatal-error-unsafe-repository-home-repon-is-owned-by-someone-else
* https://github.com/microsoft/vscode-remote-release/issues/6683
* https://github.blog/2022-04-12-git-security-vulnerability-announced/
* https://github.com/git/git/blob/1ac7422e39b0043250b026f9988d0da24cb2cb58/Documentation/config/safe.txt